update(User, Company): update crud for company and user

+++ src/main/java/vn/datk/jobhunter/service/CompanyService.java
@@ -6,16 +6,20 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import vn.datk.jobhunter.domain.Company;
+import vn.datk.jobhunter.domain.User;
 import vn.datk.jobhunter.domain.res.ResultPaginationResponse;
 import vn.datk.jobhunter.repository.CompanyRepository;
+import vn.datk.jobhunter.repository.UserRepository;
 import vn.datk.jobhunter.util.response.FormatResultPagaination;
 
+import java.util.List;
 import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
 public class CompanyService {
     private final CompanyRepository companyRepository;
+    private final UserRepository userRepository;
 
     public Company createCompany(Company company){
         return companyRepository.save(company);
@@ -25,6 +29,11 @@ public class CompanyService {
         Page<Company> companyPage = companyRepository.findAll(spec,pageable);
         ResultPaginationResponse response = FormatResultPagaination.createPaginationResponse(companyPage);
         return response;
+    }
+
+    public Company findCompanyById(Long id){
+        Optional<Company> companyOptional = this.companyRepository.findById(id);
+        return companyOptional.orElse(null);
     }
 
     public Company updateCompany(Long id, Company company){
@@ -41,6 +50,12 @@ public class CompanyService {
     }
 
     public void deleteCompany(Long id){
-        this.companyRepository.deleteById(id);
+        Optional<Company> companyOptional = this.companyRepository.findById(id);
+        if(companyOptional.isPresent()){
+            Company company = companyOptional.get();
+            List<User> users = this.userRepository.findByCompany(company);
+            this.userRepository.deleteAll(users);
+            this.companyRepository.deleteById(id);
+        }
     }
 }
